+++ src/async-utils/worker.ts
@@ -8,12 +8,12 @@ import type { SynckitOptions } from "synckit";
 export function getWorkerOptions(): SynckitOptions | undefined {
   if(env.NODE_ENV === "test"){
     return {
-      timeout: 10_000,
+      timeout: 30_000,
       tsRunner: TsRunner.OXC
     };
   } else {
     return {
-      timeout: 10_000
+      timeout: 30_000
     };
   }
 }