+++ src/rules/enforce-consistent-class-order.ts
@@ -1,12 +1,25 @@
-import { description, literal, object, optional, pipe, union } from "valibot";
+import {
+  boolean,
+  description,
+  literal,
+  object,
+  optional,
+  pipe,
+  union
+} from "valibot";
 
 import { createGetClassOrder, getClassOrder } from "better-tailwindcss:tailwindcss/class-order.js";
+import {
+  createGetCustomComponentClasses,
+  getCustomComponentClasses
+} from "better-tailwindcss:tailwindcss/custom-component-classes.js";
 import { createGetDissectedClasses, getDissectedClasses } from "better-tailwindcss:tailwindcss/dissect-classes.js";
 import { async } from "better-tailwindcss:utils/context.js";
 import { escapeNestedQuotes } from "better-tailwindcss:utils/quotes.js";
 import { createRule } from "better-tailwindcss:utils/rule.js";
 import { display, splitClasses, splitWhitespaces } from "better-tailwindcss:utils/utils.js";
 
+import type { DissectedClass } from "better-tailwindcss:tailwindcss/dissect-classes.js";
 import type { Warning } from "better-tailwindcss:types/async.js";
 import type { Context } from "better-tailwindcss:types/rule.js";
 
@@ -24,23 +37,73 @@ export const enforceConsistentClassOrder = createRule({
   },
 
   schema: object({
+    componentClassOrder: optional(
+      pipe(
+        union([
+          literal("asc"),
+          literal("desc"),
+          literal("preserve")
+        ]),
+        description("Defines how component classes should be ordered among themselves.")
+      ),
+      "preserve"
+    ),
+    componentClassPosition: optional(
+      pipe(
+        union([
+          literal("start"),
+          literal("end")
+        ]),
+        description("Defines where component classes should be placed in relation to the whole string literal.")
+      ),
+      "start"
+    ),
+    detectComponentClasses: optional(
+      pipe(
+        boolean(),
+        description("Whether to automatically detect custom component classes from the tailwindcss config.")
+      ),
+      false
+    ),
     order: optional(
       pipe(
         union([
           literal("asc"),
           literal("desc"),
           literal("official"),
-          literal("improved")
+          literal("strict")
         ]),
         description("The algorithm to use when sorting classes.")
       ),
-      "improved"
+      "official"
+    ),
+    unknownClassOrder: optional(
+      pipe(
+        union([
+          literal("asc"),
+          literal("desc"),
+          literal("preserve")
+        ]),
+        description("Defines how component classes should be ordered among themselves.")
+      ),
+      "preserve"
+    ),
+    unknownClassPosition: optional(
+      pipe(
+        union([
+          literal("start"),
+          literal("end")
+        ]),
+        description("Defines where component classes should be placed in relation to the whole string literal.")
+      ),
+      "start"
     )
   }),
 
-  initialize: (ctx: Context) => {
+  initialize: ctx => {
     createGetClassOrder(ctx);
     createGetDissectedClasses(ctx);
+    createGetCustomComponentClasses(ctx);
   },
 
   lintLiterals: (ctx, literals) => {
@@ -109,7 +172,7 @@ export const enforceConsistentClassOrder = createRule({
 
 function sortClassNames(ctx: Context<typeof enforceConsistentClassOrder>, classes: string[]): [classes: string[], warnings?: (Warning | undefined)[]] {
 
-  const { order } = ctx.options;
+  const { componentClassOrder, componentClassPosition, order, unknownClassOrder, unknownClassPosition } = ctx.options;
 
   if(order === "asc"){
     return [classes.toSorted((a, b) => a.localeCompare(b))];
@@ -120,14 +183,46 @@ function sortClassNames(ctx: Context<typeof enforceConsistentClassOrder>, classe
   }
 
   const { classOrder, warnings } = getClassOrder(async(ctx), classes);
-  const { dissectedClasses } = getDissectedClasses(async(ctx), classes);
+  const { customComponentClasses } = getCustomComponentClasses(async(ctx));
 
   const officiallySortedClasses = classOrder
-    .toSorted(([, a], [, z]) => {
-      if(a === z){ return 0; }
-      if(a === null){ return -1; }
-      if(z === null){ return +1; }
-      return +(a - z > 0n) - +(a - z < 0n);
+    .toSorted((a, b) => {
+      const [classA, aIndex] = a;
+      const [classB, bIndex] = b;
+
+      const componentClassSorting = getCustomOrder(
+        componentClassPosition,
+        componentClassOrder,
+        classA,
+        classB,
+        className => customComponentClasses.includes(className)
+      );
+
+      if(componentClassSorting !== undefined){
+        return componentClassSorting;
+      }
+
+      const unknownClassSorting = getCustomOrder(
+        unknownClassPosition,
+        unknownClassOrder,
+        classA,
+        classB,
+        className => {
+          return (
+            (classA === className && aIndex === null || classB === className && bIndex === null) &&
+            !customComponentClasses.includes(className)
+          );
+        }
+      );
+
+      if(unknownClassSorting !== undefined){
+        return unknownClassSorting;
+      }
+
+      if(aIndex === bIndex){ return 0; }
+      if(aIndex === null){ return -1; }
+      if(bIndex === null){ return +1; }
+      return +(aIndex - bIndex > 0n) - +(aIndex - bIndex < 0n);
     })
     .map(([className]) => className);
 
@@ -135,14 +230,132 @@ function sortClassNames(ctx: Context<typeof enforceConsistentClassOrder>, classe
     return [officiallySortedClasses, warnings];
   }
 
-  const groupedByVariant = new Map<string, string[]>();
+  const { dissectedClasses } = getDissectedClasses(async(ctx), classes);
 
-  for(const className of officiallySortedClasses){
-    const dissectedClass = dissectedClasses.find(dissectedClass => dissectedClass.className === className);
-    const variants = dissectedClass?.variants.join(":") ?? "";
-    groupedByVariant.set(variants, [...groupedByVariant.get(variants) ?? [], className]);
+  const variantMap: VariantMap = {};
+
+  for(const dissectedClass of dissectedClasses){
+    dissectedClass.variants.unshift("");
+
+    for(let v = 0, variantMapLevel = variantMap; v < dissectedClass.variants.length; v++){
+      const isLastVariant = v === dissectedClass.variants.length - 1;
+
+      variantMapLevel[dissectedClass.variants[v]] ??= {
+        dissectedClasses: [],
+        nested: {}
+      };
+
+      if(isLastVariant){
+        variantMapLevel[dissectedClass.variants[v]].dissectedClasses.push(dissectedClass);
+        continue;
+      }
+
+      variantMapLevel = variantMapLevel[dissectedClass.variants[v]].nested;
+
+    }
   }
 
-  return [Array.from(groupedByVariant.values()).flat(), warnings];
+  const strictOrder = getStrictOrder(variantMap);
 
+  return [strictOrder, warnings];
+
+}
+
+
+type VariantMap = {
+  [variant: string]: {
+    dissectedClasses: DissectedClass[];
+    nested: VariantMap;
+  };
+};
+
+function getStrictOrder(variantMap: VariantMap): string[] {
+  const orderedClasses: string[] = [];
+
+  const orderedVariants = Object.keys(variantMap).sort((a, b) => {
+    const aIsArbitrary = isArbitrary(a);
+    const bIsArbitrary = isArbitrary(b);
+
+    // sort arbitrary variants last
+    if(aIsArbitrary && !bIsArbitrary){ return +1; }
+    if(!aIsArbitrary && bIsArbitrary){ return -1; }
+
+    return 0;
+  });
+
+  for(let v = 0; v < orderedVariants.length; v++){
+    const variant = orderedVariants[v];
+    const nextVariant = orderedVariants[v + 1];
+
+    const variantIsArbitrary = isArbitrary(variant);
+    const nextVariantIsArbitrary = isArbitrary(nextVariant);
+
+    const { dissectedClasses, nested } = variantMap[variant];
+
+    orderedClasses.push(...dissectedClasses.map(dissectedClass => dissectedClass.className));
+
+    if(dissectedClasses.length > 0 || !variantIsArbitrary && nextVariantIsArbitrary){
+      orderedClasses.push(...getStrictOrder(nested));
+    }
+  }
+
+  for(let v = 0; v < orderedVariants.length; v++){
+    const variant = orderedVariants[v];
+    const nextVariant = orderedVariants[v + 1];
+
+    const variantIsArbitrary = isArbitrary(variant);
+    const nextVariantIsArbitrary = isArbitrary(nextVariant);
+
+    const { dissectedClasses, nested } = variantMap[variant];
+
+    if(!(dissectedClasses.length > 0 || !variantIsArbitrary && nextVariantIsArbitrary)){
+      orderedClasses.push(...getStrictOrder(nested));
+    }
+  }
+
+  return orderedClasses;
+
+}
+
+function getCustomOrder(position: "end" | "start", order: "asc" | "desc" | "preserve", classA: string, classB: string, isCustomClass: (className: string) => boolean): number | undefined {
+  const aIsCustomClass = isCustomClass(classA);
+  const bIsCustomClass = isCustomClass(classB);
+
+  if(position === "start"){
+    if(aIsCustomClass && !bIsCustomClass){ return -1; }
+    if(!aIsCustomClass && bIsCustomClass){ return +1; }
+
+    if(aIsCustomClass && bIsCustomClass){
+      if(order === "asc"){
+        return classA.localeCompare(classB);
+      }
+      if(order === "desc"){
+        return classB.localeCompare(classA);
+      }
+      return 0;
+    }
+  }
+  if(position === "end"){
+    if(aIsCustomClass && !bIsCustomClass){ return +1; }
+    if(!aIsCustomClass && bIsCustomClass){ return -1; }
+
+    if(aIsCustomClass && bIsCustomClass){
+      if(order === "asc"){
+        return classA.localeCompare(classB);
+      }
+      if(order === "desc"){
+        return classB.localeCompare(classA);
+      }
+      return 0;
+    }
+  }
+
+}
+
+function isArbitrary(variant?: string): boolean {
+  if(!variant){
+    return false;
+  }
+
+  return variant.includes("[") && variant.includes("]");
 }